Update IInAppBillingService.aidl from GitHub
It should be identical to the version at
https://raw.githubusercontent.com/googlesamples/android-play-billing/9dbfdfd15cd3153dae03375b1e66f9fbf206debf/TrivialDrive/app/src/main/aidl/com/android/vending/billing/IInAppBillingService.aidl

I'm hopeful this will make the call to getPurchaseHistory() Just Work,
as that's listed in Google's official Android docs
(https://developer.android.com/google/play/billing/billing_reference.html#getPurchaseHistory)
but wasn't in the last committed version of this AIDL file.

I guess we'll see whether this was a mistake.