Fix inAppPurchase.getProducts() method on Android
The previous implementation fails when there are more than twenty
products and the user owns some of them.

This is just a copy of this PR that upstream has never accepted:

https://github.com/AlexDisler/cordova-plugin-inapppurchase/pull/57

Note that this undoes the conceptually-wrong change that *was* accepted
to deal with this problem (but doesn't):

https://github.com/AlexDisler/cordova-plugin-inapppurchase/pull/94

I've worked through this patch line-by-line and it seems like a
reasonable solution to me. Granted, it's been a few years since I wrote
Java, but conceptually it seems right - handle the Google API's demand
for chunking internally to the querySkuDetails() method, so nothing else
has to know about it.